Reshuffle main function to allow for programmatic usage
NEW: define (imperfect) programmatic usage of rdiff-backup, see FAQ for details, closes #703
